# backend_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mine Worker Safety Management System API",
    "description": "Backend API for Role-Based Access Control, Shift Monitoring, Hazard Reporting, Live Tracking, and SOS Alerts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Users",
        "operationId": "get_users_api_users__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "type": "array",
                  "title": "Response Get Users Api Users  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User By Id",
        "operationId": "get_user_by_id_api_users__user_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User",
        "operationId": "update_user_api_users__user_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/status": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Toggle User Status",
        "operationId": "toggle_user_status_api_users__user_id__status_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workers/": {
      "post": {
        "tags": [
          "Workers"
        ],
        "summary": "Create Worker",
        "operationId": "create_worker_api_workers__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkerPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/workers/profiles": {
      "get": {
        "tags": [
          "Workers"
        ],
        "summary": "List Worker Profiles",
        "operationId": "list_worker_profiles_api_workers_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkerProfileOut"
                  },
                  "type": "array",
                  "title": "Response List Worker Profiles Api Workers Profiles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/workers/{worker_id}/profile": {
      "get": {
        "tags": [
          "Workers"
        ],
        "summary": "Get Worker Profile",
        "operationId": "get_worker_profile_api_workers__worker_id__profile_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workers"
        ],
        "summary": "Update Worker Profile",
        "operationId": "update_worker_profile_api_workers__worker_id__profile_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Worker Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkerPayload": {
        "properties": {
          "employee_id": {
            "type": "string",
            "title": "Employee Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "emergency_contact_name": {
            "type": "string",
            "title": "Emergency Contact Name"
          },
          "emergency_contact_number": {
            "type": "string",
            "title": "Emergency Contact Number"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "blood_group": {
            "type": "string",
            "title": "Blood Group"
          },
          "medical_conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Conditions"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "mine_location": {
            "type": "string",
            "title": "Mine Location"
          },
          "designation": {
            "type": "string",
            "title": "Designation"
          },
          "joining_date": {
            "type": "string",
            "format": "date",
            "title": "Joining Date"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "employee_id",
          "full_name",
          "age",
          "gender",
          "phone_number",
          "emergency_contact_name",
          "emergency_contact_number",
          "address",
          "blood_group",
          "department",
          "mine_location",
          "designation",
          "joining_date",
          "username",
          "email",
          "password"
        ],
        "title": "CreateWorkerPayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "worker"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkerProfileOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "UserOut"
      },
      "UserUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkerProfileOut": {
        "properties": {
          "employee_id": {
            "type": "string",
            "title": "Employee Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "emergency_contact_name": {
            "type": "string",
            "title": "Emergency Contact Name"
          },
          "emergency_contact_number": {
            "type": "string",
            "title": "Emergency Contact Number"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "blood_group": {
            "type": "string",
            "title": "Blood Group"
          },
          "medical_conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Conditions"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "mine_location": {
            "type": "string",
            "title": "Mine Location"
          },
          "designation": {
            "type": "string",
            "title": "Designation"
          },
          "joining_date": {
            "type": "string",
            "format": "date",
            "title": "Joining Date"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "safety_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Safety Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "employee_id",
          "full_name",
          "age",
          "gender",
          "phone_number",
          "emergency_contact_name",
          "emergency_contact_number",
          "address",
          "blood_group",
          "department",
          "mine_location",
          "designation",
          "joining_date",
          "id",
          "user_id",
          "safety_score",
          "created_at",
          "updated_at"
        ],
        "title": "WorkerProfileOut"
      },
      "WorkerProfileUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Number"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "blood_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Group"
          },
          "medical_conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medical Conditions"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "mine_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mine Location"
          },
          "designation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Designation"
          },
          "joining_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joining Date"
          }
        },
        "type": "object",
        "title": "WorkerProfileUpdate"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}